backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9

# Cache & Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Config
python-decouple==3.8
gunicorn==21.2.0

# Kafka
confluent-kafka==2.3.0

# Reports
reportlab==4.0.8
openpyxl==3.1.2
xlsxwriter==3.1.9

# AWS S3 (optional)
boto3==1.34.0

# Utils
Pillow==10.2.0
